+++ docs/changelog.md
@@ -4,6 +4,12 @@ Dates using [ISO 8601 Format](http://www.iso.org/iso/iso8601) (YYYY-MM-DD).
 
 Check whether the latest build has passed: [Builds | Read the Docs](https://readthedocs.org/projects/wireshell/builds/)
 
+**1.0.3** (2018-11-16)
+
+- Adds `template:tag` command
+- Adds `template:info` command
+- Adds `field:rename` command
+
 **1.0.2** (2017-03-10)
 
 - Fixes parse error for usage with php below version 7